refactor(walkscore): move WalkScoreService to walkscore package
