+++ src/main/java/com/kakao/termproject/walkscore/service/WalkScoreService.java
@@ -1,10 +1,9 @@
-package com.kakao.termproject.weather.service;
+package com.kakao.termproject.walkscore.service;
 
 import com.kakao.termproject.pet.domain.Pet;
 import com.kakao.termproject.pet.service.PetService;
 import com.kakao.termproject.walk.domain.Walk;
 import com.kakao.termproject.walk.service.WalkService;
-import com.kakao.termproject.walkscore.WalkScoreCalculator;
 import com.kakao.termproject.weather.dto.WeatherDetailInternal;
 import com.kakao.termproject.weather.dto.WeatherResponse;
 import com.kakao.termproject.weather.dto.WeatherResponse.HourlyForecast;
